% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_beta}
\alias{coeff_beta}
\title{Function for Beta Estimator}
\usage{
coeff_beta(Dependent, Independent)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (or outcome variable).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}
}
\value{
A \code{numeric}.
 \describe{
     \item{beta}{Estimated coefficients, Linear Regression Model.}
}
}
\description{
We create a function of the OLS estimator of the coefficients called coeff_beta.
}
\examples{
data("USArrests")
coeff_beta(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
