% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_ci}
\alias{coeff_ci}
\title{Linear Model Function-group3}
\usage{
coeff_ci(Dependent, Independent, alpha = 0.05)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (or outcome variable).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}

\item{alpha}{A \code{numeric} (double) that sets the alpha coefficient to be used. Has to be between 0 and 1.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{ci.beta}{confidence interval for the estimated beta.}
}
}
\description{
We create a function of the OLS estimator of the coefficients called beta_hats.
}
\examples{
data("USArrests")
coeff_ci(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
