% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_ftest}
\alias{coeff_ftest}
\title{Function for F-test -group3}
\usage{
coeff_ftest(Dependent, Independent)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{f-stat}{value for f_star}
     \item{DFM}{Degrees of freedom}
     \item{DFE}{p -1}
     \item{f_prob}{f's p-value}
}
}
\description{
Function that calculates the F-test
}
\examples{
data("USArrests")
coeff_ftest(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
