% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_mallows_Cp}
\alias{coeff_mallows_Cp}
\title{Function for Mallows Cp-group3}
\usage{
coeff_mallows_Cp(Dependent, Independent)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{mallows}{value for Mallows Cp}
}
}
\description{
Function that calculates Mallows Cp
}
\examples{
data("USArrests")
coeff_mallows_Cp(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
