% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_plot}
\alias{coeff_plot}
\title{Plot function for package-group3}
\usage{
coeff_plot(Dependent, Independent)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}
}
\value{
A \code{list} containing the following attributes:
}
\description{
Function that creates the 3 required plots
}
\examples{
data("USArrests")
coeff_plot(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
