% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liner_model-G3.R
\name{coeff_rsquared}
\alias{coeff_rsquared}
\title{Function for R Squared}
\usage{
coeff_rsquared(Dependent, Independent)
}
\arguments{
\item{Dependent}{A \code{vector} with the values for the dependent variable (also called outcome).}

\item{Independent}{A \code{matrix} with the values for the independent variable.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{r_squared}{gets the r squared value}
}
}
\description{
Function that calculates the R squared value
}
\examples{
data("USArrests")
coeff_rsquared(USArrests$Murder,USArrests$UrbanPop)
}
\author{
John M. Musah & Alan R. Galloway
}
